# 05_openapi_final.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: "1.0.0"
  summary: Reveal specification for the Kaveri Stays API assignment.
  description: |
    # Read this before you read the paths

    This is ONE reasonable design. It is not the only one. Several of the choices
    below are genuinely arguable and the Interlude expects you to argue with them.
    A few are not arguable, and those are marked.

    ## Conventions this specification commits to

    **Dates.** Every date is ISO 8601 `YYYY-MM-DD`. A stay is the half-open interval
    `[check_in, check_out)`. The check-out date is NOT part of the stay. This is what
    makes same-day turnover work (rule 5), and it is NOT arguable.

    **Money.** Every amount is a decimal string, never a JSON number. `"4500.00"`,
    not `4500.0`. Floats lose money.

    **Empty collections.** A list endpoint with no matching rows returns 200 and an
    empty `items` array. It never returns 404. 404 means the container does not exist,
    not that it is empty.

    **401 versus 403.** 401 means the server does not know who you are: no token, an
    expired token, a forged token. 403 means it knows exactly who you are and you may
    not do this. This distinction is NOT arguable.

    **Request bodies.** Every request schema sets `additionalProperties: false`. A
    field you did not declare is rejected rather than ignored. This is what defeats
    attacks 8.2 and 8.8 structurally rather than by remembering to check.

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /properties/{property_id}:
    get:
      tags: [properties]
      operationId: getProperty
      summary: One property.
      security: []
      x-roles: [public]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
      responses:
        '200':
          description: The property.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        '404':
          $ref: '#/components/responses/NotFound'
  /properties/{property_id}/rooms:
    get:
      tags: [properties]
      operationId: listRooms
      summary: Full room inventory for a property.
      description: |
        Every room, including rooms that have never been booked. Task 4.7. If your
        implementation joins through `bookings`, never-booked rooms vanish and several
        of your Stage 4 queries from the database assignment stop agreeing with this.
      x-roles: [staff, manager, owner]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Rooms.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomPage' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /properties/{property_id}/availability:
    get:
      tags: [properties]
      operationId: getAvailability
      summary: Rooms free for a whole date range.
      description: |
        Task 4.1, and the query the booking desk runs all day. Because a stay is
        `[check_in, check_out)`, a room whose previous guest departs on `from` IS
        available. Cancelled and no-show bookings do not occupy a room.
      security: []
      x-roles: [public]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
        - name: from
          in: query
          required: true
          description: First night of the stay, inclusive.
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          description: Departure date, exclusive. Must be strictly after `from`.
          schema: { type: string, format: date }
        - name: room_type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/RoomTypeName' }
      responses:
        '200':
          description: Available rooms, with the rate for the requested range.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /reports/occupancy:
    get:
      tags: [reports]
      operationId: reportOccupancy
      summary: Occupancy percentage per property per month.
      description: |
        A manager omitting `property_id` gets their own property. A manager naming a
        property that is not theirs gets 403, not an empty result. This is attack 8.7.

        The denominator is room-nights available, counting every room in the property
        whether or not it was ever booked.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: Occupancy rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/OccupancyRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /reports/adr:
    get:
      tags: [reports]
      operationId: reportAdr
      summary: Average daily rate per property per month.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: ADR rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RateMetricRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/revpar:
    get:
      tags: [reports]
      operationId: reportRevpar
      summary: Revenue per available room, per property per month.
      x-roles: [manager, owner]
      parameters:
        - $ref: '#/components/parameters/ReportPropertyId'
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: RevPAR rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RateMetricRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/revenue:
    get:
      tags: [reports]
      operationId: reportRevenue
      summary: Revenue per property per month, across all properties.
      description: |
        Owner only. A manager receives 403 even for their own property, because this
        endpoint's shape is cross-property by definition. Managers use the three
        endpoints above.
      x-roles: [owner]
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
      responses:
        '200':
          description: Revenue rows, with a grand total.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, grand_total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RevenueRow' }
                  grand_total: { type: string, examples: ["4820500.00"] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /guests:
    get:
      tags: [guests]
      operationId: listGuests
      summary: Guest records.
      description: |
        Staff and above only. There is deliberately no endpoint through which a guest
        can list, search or count other guests. Task 4.8.
      x-roles: [staff, manager, owner]
      parameters:
        - name: email
          in: query
          description: Exact match, case-insensitive. Not a prefix search.
          schema: { type: string, format: email }
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Guests.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /guests/{guest_id}:
    get:
      tags: [guests]
      operationId: getGuest
      summary: One guest record.
      x-roles: [staff, manager, owner]
      parameters:
        - name: guest_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The guest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Guest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token in `Authorization: Bearer <token>`. Short lived. Obtain a new one
        from `/auth/refresh` rather than by logging in again.
  parameters:
    PropertyIdPath:
      name: property_id
      in: path
      required: true
      schema: { type: integer }
...
    LimitParam:
      name: limit
      in: query
      description: Page size. Capped server-side; an unbounded list endpoint is a defect.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    ReportPropertyId:
      name: property_id
      in: query
      description: Omit to mean "the property I manage". Owners may name any property.
      schema: { type: integer }
    ReportFrom:
      name: from
      in: query
      required: true
      schema: { type: string, format: date }
    ReportTo:
      name: to
      in: query
      required: true
      schema: { type: string, format: date }
  responses:
    Unauthorized:
      description: No token, an expired token, or a token that does not verify.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: unauthenticated, message: Authentication required. }
    Forbidden:
      description: Authenticated, and not permitted to do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: forbidden, message: You do not have access to this resource. }
    NotFound:
      description: No such resource, or none you are permitted to know about.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: not_found, message: Not found. }
...
    Unprocessable:
      description: |
        The request parsed and a field is unacceptable. Includes guest counts above
        the room type maximum, check-out on or before check-in, ratings outside 1 to 5,
        and any field not declared in the request schema.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              detail:
                - field: guests
                  reason: Exceeds maximum occupancy of 2 for room type Standard.
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      description: |
        Every non-2xx response in this API uses this shape. One handler produces it.
        If you are constructing this dict in more than one module, consolidate.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable, snake_case. Clients branch on this.
              examples: [room_unavailable, illegal_transition, validation_failed]
            message:
              type: string
              description: Human-readable. Safe to show a guest. Never a raw Postgres message.
            detail:
              description: Optional structured context. Never leaks table or column names.
            request_id:
              type: string
              description: Correlates this response with your logs.
    PageMeta:
      type: object
      additionalProperties: false
      required: [limit, offset, total]
      properties:
        limit:  { type: integer }
        offset: { type: integer }
        total:  { type: integer, description: Total matching rows, ignoring pagination. }
...
    RoomTypeName:
      type: string
      enum: [Standard, Deluxe, Suite]
...
    Property:
      type: object
      additionalProperties: false
      required: [id, name, city]
      properties:
        id:    { type: integer }
        name:  { type: string, examples: ["Kaveri Riverside"] }
        city:  { type: string, examples: ["Coorg"] }
        stars: { type: integer, minimum: 1, maximum: 5 }
    RoomType:
      type: object
      additionalProperties: false
      required: [name, max_occupancy]
      properties:
        name:          { $ref: '#/components/schemas/RoomTypeName' }
        max_occupancy: { type: integer, minimum: 1 }
    Room:
      type: object
      additionalProperties: false
      required: [id, property_id, room_number, room_type]
      properties:
        id:          { type: integer }
        property_id: { type: integer }
        room_number: { type: string, description: Unique within a property, repeats across them. Rule 2. }
        room_type:   { $ref: '#/components/schemas/RoomType' }
    RoomPage:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Room' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    AvailableRoom:
      type: object
      additionalProperties: false
      required: [room_id, room_number, room_type, nights, total_rate]
      properties:
        room_id:     { type: integer }
        room_number: { type: string }
        room_type:   { $ref: '#/components/schemas/RoomType' }
        nights:      { type: integer }
        total_rate:
          type: string
          description: |
            Sum of the nightly rates across the requested range, resolved from
            `rate_plans`. A stay crossing two seasons is priced per night, so this is
            not `nights` multiplied by any single rate.
          examples: ["13500.00"]
    AvailabilityResponse:
      type: object
      additionalProperties: false
      required: [property_id, from, to, items]
      properties:
        property_id: { type: integer }
        from:        { type: string, format: date }
        to:          { type: string, format: date }
        items:
          type: array
          items: { $ref: '#/components/schemas/AvailableRoom' }
...
    OccupancyRow:
      type: object
      additionalProperties: false
      required: [property_id, month, room_nights_available, room_nights_sold, occupancy_pct]
      properties:
        property_id:           { type: integer }
        property_name:         { type: string }
        month:                 { type: string, examples: ["2025-12"] }
        room_nights_available: { type: integer }
        room_nights_sold:      { type: integer }
        occupancy_pct:         { type: string, examples: ["78.40"] }
    RateMetricRow:
      type: object
      additionalProperties: false
      required: [property_id, month, value]
      properties:
        property_id:   { type: integer }
        property_name: { type: string }
        month:         { type: string, examples: ["2025-12"] }
        value:         { type: string, examples: ["6420.00"] }
    RevenueRow:
      type: object
      additionalProperties: false
      required: [property_id, month, revenue]
      properties:
        property_id:   { type: integer }
        property_name: { type: string }
        month:         { type: string, examples: ["2025-12"] }
        revenue:       { type: string, examples: ["1284000.00"] }
    Guest:
      type: object
      additionalProperties: false
      required: [id, email, full_name]
      properties:
        id:         { type: integer }
        email:      { type: string, format: email }
        full_name:  { type: string }
        phone:      { type: string }
        stay_count: { type: integer }
    GuestPage:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Guest' }
        meta: { $ref: '#/components/schemas/PageMeta' }